requirements.txt
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.4
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
model-bakery==1.18.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.0
pytest-django==4.8.0
pytest-testdox==3.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
